+++ pyproject.toml
@@ -25,9 +25,9 @@ streaming_memory_agent = "datarobot_genai.dragent.plugins.streaming_memory_agent
 datarobot_otel_conventions_middleware = "datarobot_genai.dragent.plugins.datarobot_otel_conventions_middleware"
 datarobot_mcp_client = "datarobot_genai.nat.datarobot_mcp_client"
 datarobot_auth_provider = "datarobot_genai.nat.datarobot_auth_provider"
-datarobot_mem0_memory = "datarobot_genai.nat.datarobot_mem0_memory"
 datarobot_moderation_middleware = "datarobot_genai.nat.datarobot_moderation_middleware"
 datarobot_otelcollector = "datarobot_genai.dragent.plugins.datarobot_otelcollector"
+datarobot_mem0_memory = "datarobot_genai.dragent.plugins.datarobot_mem0_memory"
 llm_providers = "datarobot_genai.dragent.plugins.llm_providers"
 llm_clients = "datarobot_genai.dragent.plugins.llm_clients"
 crewai_llm_clients = "datarobot_genai.crewai.llm_clients"
